ssec/requirements.txt
tk
argon2-cffi==23.1.0
argon2-cffi-bindings=21.2.0
cffi==1.16.0
pycparser==2.22
pycryptodomex==3.11.0